+++ crayg/utils/StopWatch.h
@@ -7,6 +7,8 @@
 
 #include <chrono>
 #include <iostream>
+#define FMT_HEADER_ONLY
+#include "fmt/format.h"
 
 class StopWatch{
 public:
@@ -16,8 +18,8 @@ public:
     void end(){
         std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();
         auto microseconds = std::chrono::duration_cast<std::chrono::microseconds>(end - begin).count();
-        // todo use format library
-        std::cout << name << " took " << microseconds *0.0000006 << " seconds" << std::endl;
+
+        std::cout << fmt::format("{} took {} seconds.\n", name, microseconds *0.0000006);
     };
 private:
     std::chrono::steady_clock::time_point begin;